Admin lte 3 mastering and locale completed

+++ app/Http/Controllers/CategoryController.php
@@ -45,7 +45,7 @@ class CategoryController extends Controller
 
         Category::create($request->all());
 
-        return redirect('/categories')->with('status', 'Created successful');
+        return redirect('/categories')->with('success', 'Created successful');
 
     }
 
@@ -86,19 +86,23 @@ class CategoryController extends Controller
         ]);
 
         $category->update($request->all());
-        return redirect('/categories')->with('status', 'Updated successful');
+        return redirect('/categories')->with('success', 'Updated successful');
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Category  $category
-     * @return \Illuminate\Http\Response
+     * @param Category $category
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Exception
      */
     public function destroy(Category $category)
     {
-       $category->delete();
-        return redirect('/categories')->with('status', 'Deleted successful');
+        if($category->posts->count()  === 0){
+            $category->delete();
+        return redirect('/categories')->with('success', 'Deleted successful');
+        }else {
+            return redirect('/categories')->with('error', "Unable to delete, it's being used by another program!!");
+        }
+
     }
 
     /**
@@ -110,7 +114,7 @@ class CategoryController extends Controller
             'status' => !$category->status
         ]);
 
-        return redirect('/categories')->with('status', 'Updated successful');
+        return redirect('/categories')->with('success', 'Updated successful');
 
     }
 
@@ -127,11 +131,9 @@ class CategoryController extends Controller
     public function restoreCategory($id)
     {
         $category = Category::onlyTrashed()->findOrFail($id);
-//        dd($category);
-
         $category->restore();
 
-        return redirect('/categories/trash')->with('status', 'Restored successful');
+        return redirect('/categories/trash')->with('success', 'Restored successful');
 
     }
 }
